gui: refactored remaining windows
