+++ src/platform/windows/gui/help/about.h
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace gui::help {
+class About {
+    void aboutWindow();
+
+   public:
+    bool aboutWindowOpen = false;
+
+    void displayWindows();
+};
+}  // namespace gui::help
