+++ src/platform/windows/gui/options/options.h
