--- src/platform/windows/gui/tools.h
@@ -1,7 +1,0 @@
-#include <imgui.h>
-#include <string>
-#include <vector>
-
-namespace gui::tools {
-void Columns(const std::vector<std::string>& list, int col_num = 4);
-};
